New version of file debian/README.source.
Because of changes to the packages in Debian (and presumably
therefore Ubuntu) and at the recommdation of the author, need
to change over to using the various non-sbbs libraries provided
with the synchronet v3.14a archives.   Because of that, needed
to change the discription of how the source archive is created,
and this updated README.source file is the result.